feat-2308: move and update legacy tests
